+++ tests/testthat/test-burrIII2.R
@@ -12,102 +12,162 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-test_that("dburrIII2", {
+test_that("dburrIII2 extremes", {
+  lifecycle::expect_deprecated(dburrIII2(1))
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
+
   expect_identical(dburrIII2(numeric(0)), numeric(0))
   expect_identical(dburrIII2(NA), NA_real_)
-  expect_equal(dburrIII2(c(0, 1, Inf, NaN, -1)), c(0, 0.196611933241482, 0, NA, 0))
+  expect_identical(dburrIII2(NaN), NaN)
+  expect_identical(dburrIII2(0), 0)
+  expect_equal(dburrIII2(1), 0.25)
+  expect_equal(dburrIII2(1, log = TRUE), log(dburrIII2(1)))
+  expect_identical(dburrIII2(0), 0)
+  expect_identical(dburrIII2(-Inf), 0)
+  expect_identical(dburrIII2(Inf), 0)
+  expect_identical(dburrIII2(c(NA, NaN, 0, Inf, -Inf)), 
+                   c(dburrIII2(NA), dburrIII2(NaN), dburrIII2(0), dburrIII2(Inf), dburrIII2(-Inf)))
+  expect_equal(dburrIII2(1:2, 1:2, 3:4), 
+               c(dburrIII2(1, 1, 3), dburrIII2(2, 2, 4)))
+  expect_equal(dburrIII2(1:2, c(1, NA), 3:4), 
+               c(dburrIII2(1, 1, 3), NA))
+})
+
+test_that("pburrIII2 extremes", {
+  lifecycle::expect_deprecated(pburrIII2(1))
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
+
+  expect_identical(pburrIII2(numeric(0)), numeric(0))
+  expect_identical(pburrIII2(NA), NA_real_)
+  expect_identical(pburrIII2(NaN), NaN)
+  expect_identical(pburrIII2(0), 0)
+  expect_equal(pburrIII2(1), 0.5)
+  expect_equal(pburrIII2(1, log.p = TRUE), log(pburrIII2(1)))
+  expect_equal(pburrIII2(1, lower.tail = FALSE), 1 - 0.5)
+  expect_equal(pburrIII2(1, lower.tail = FALSE, log.p = TRUE), log(1 - 0.5))
+  expect_identical(pburrIII2(0), 0)
+  expect_identical(pburrIII2(-Inf), 0)
+  expect_identical(pburrIII2(Inf), 1)
+  expect_identical(pburrIII2(c(NA, NaN, 0, Inf, -Inf)), 
+                   c(pburrIII2(NA), pburrIII2(NaN), pburrIII2(0), pburrIII2(Inf), pburrIII2(-Inf)))
+  expect_equal(pburrIII2(1:2, 1:2, 3:4), 
+               c(pburrIII2(1, 1, 3), pburrIII2(2, 2, 4)))
+  expect_equal(pburrIII2(1:2, c(1, NA), 3:4), 
+               c(pburrIII2(1, 1, 3), NA))
+})
+
+test_that("qburrIII2 extremes", {
+  lifecycle::expect_deprecated(qburrIII2(1))
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
+
+  expect_identical(qburrIII2(numeric(0)), numeric(0))
+  expect_identical(qburrIII2(NA), NA_real_)
+  expect_identical(qburrIII2(NaN), NaN)
+  expect_identical(qburrIII2(0), 0)
+  expect_identical(qburrIII2(1), Inf)
+  expect_equal(qburrIII2(0.75), 3)
+  expect_equal(qburrIII2(0.75, log.p = TRUE), NaN)
+  expect_equal(qburrIII2(log(0.75), log.p = TRUE), qburrIII2(0.75))
+  expect_equal(qburrIII2(0.75, lower.tail = FALSE), qburrIII2(0.25))
+  expect_equal(qburrIII2(log(0.75), lower.tail = FALSE, log.p = TRUE), qburrIII2(0.25))
+  expect_identical(qburrIII2(0), 0)
+  expect_identical(qburrIII2(-Inf), NaN)
+  expect_identical(qburrIII2(Inf), NaN)
+  expect_identical(qburrIII2(c(NA, NaN, 0, Inf, -Inf)), 
+                   c(qburrIII2(NA), qburrIII2(NaN), qburrIII2(0), qburrIII2(Inf), qburrIII2(-Inf)))
+  expect_equal(qburrIII2(1:2, 1:2, 3:4), 
+               c(qburrIII2(1, 1, 3), qburrIII2(2, 2, 4)))
+  expect_equal(qburrIII2(1:2, c(1, NA), 3:4), 
+               c(qburrIII2(1, 1, 3), NA))
+  expect_equal(qburrIII2(pburrIII2(c(0, 0.1, 0.5, 0.9, 1))), c(0, 0.1, 0.5, 0.9, 1))
+})
+
+test_that("rburrIII2 extremes", {
+  lifecycle::expect_deprecated(rburrIII2(1))
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
+
+  expect_identical(rburrIII2(numeric(0)), numeric(0))
+  expect_error(rburrIII2(NA))
+  expect_identical(rburrIII2(0), numeric(0))
+  set.seed(42)
+  expect_equal(rburrIII2(1), 10.7379218085407)
+  set.seed(42)
+  expect_equal(rburrIII2(1.9), 10.7379218085407)
+  set.seed(42)
+  expect_equal(rburrIII2(2), c(10.7379218085407, 14.8920392236127))
+  set.seed(42)
+  expect_equal(rburrIII2(3:4), c(10.7379218085407, 14.8920392236127))
+  expect_equal(rburrIII2(0), numeric(0))
+  expect_error(rburrIII2(1, shape1 = 1:2))
+  expect_error(rburrIII2(1, shape2 = 1:2))
+  expect_identical(rburrIII2(1, NA), NA_real_)
+})
+
+test_that("dburrIII2", {
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
 
   expect_equal(
-    dburrIII2(c(31, 15, 32, 32, 642, 778, 187, 12), lscale = 0),
+    dburrIII2(c(31, 15, 32, 32, 642, 778, 187, 12)),
     c(
       0.0009765625, 0.00390625, 0.000918273645546373, 0.000918273645546373,
       2.41867799897932e-06, 1.64787810975198e-06, 2.82933454051607e-05,
       0.00591715976331361
     )
   )
-  expect_equal(
-    dburrIII2(c(31, 15, 32, 32, 642, 778, 187, 12), lscale = 0, log = TRUE),
-    log(c(
-      0.0009765625, 0.00390625, 0.000918273645546373, 0.000918273645546373,
-      2.41867799897932e-06, 1.64787810975198e-06, 2.82933454051607e-05,
-      0.00591715976331361
-    ))
-  )
 })
 
-test_that("fit burrIII2", {
+test_that("fit dburrIII2", {
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
   data <- data.frame(Conc = c(31, 15, 32, 32, 642, 778, 187, 12))
 
-  dist <- ssdtools:::ssd_fit_dist(data, dist = "burrIII2")
+  dist <- ssd_fit_dist(data, dist = "burrIII2")
 
   expect_true(is.fitdist(dist))
-  expect_equal(coef(dist), c(lshape = 0.101242679105766, lscale = -3.92364881206751))
+  expect_equal(coef(dist), c(locationlog = 4.07989187488155, scalelog = 0.937136717088929
+  ))
 
   data$Conc <- data$Conc / 1000
 
-  dist <- ssdtools:::ssd_fit_dist(data, dist = "burrIII2")
+  dist <- ssd_fit_dist(data, dist = "burrIII2")
 
   expect_true(is.fitdist(dist))
-  expect_equal(coef(dist), c(lshape = 0.101250994107126, lscale = 2.98432920414956))
+  expect_equal(coef(dist), c(locationlog = -2.82742464158491, scalelog = 0.937137287983746
+  ))
 })
 
-test_that("fit burrIII2 cis", {
-  dist <- ssdtools:::ssd_fit_dist(ssdtools::boron_data, dist = "burrIII2")
+test_that("fit dburrIII2 cis", {
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
+  dist <- ssd_fit_dist(ssdtools::boron_data, dist = "burrIII2")
 
   set.seed(77)
   expect_equal(
     as.data.frame(ssd_hc(dist, ci = TRUE, nboot = 10)),
-    structure(list(
-      percent = 5, est = 1.58920212463066, se = 0.561779293276257,
-      lcl = 0.933413683053849, ucl = 2.4171610278875, dist = "burrIII2"
-    ), class = "data.frame", row.names = c(
-      NA,
-      -1L
-    ))
+    structure(list(percent = 5, est = 1.56256632555312, se = 0.533080817076427, 
+                   lcl = 1.02028787432084, ucl = 2.4100026087857, dist = "burrIII2"), row.names = c(NA, 
+                                                                                                    -1L), class = "data.frame")
   )
   set.seed(77)
   expect_equal(
     as.data.frame(ssd_hp(dist, conc = 2, ci = TRUE, nboot = 10)),
-    structure(list(
-      conc = 2, est = 6.84079666045762, se = 2.8457114994266,
-      lcl = 3.57247087396272, ucl = 11.2796053723563, dist = "burrIII2"
-    ), class = "data.frame", row.names = c(
-      NA,
-      -1L
-    ))
+    structure(list(conc = 2, est = 6.8431214411277, se = 2.53471375726922, 
+                   lcl = 3.73102816675466, ucl = 10.3929969316288, dist = "burrIII2"), row.names = c(NA, 
+                                                                                                     -1L), class = "data.frame")
   )
 })
 
-test_that("qburrIII2", {
-  expect_identical(qburrIII2(numeric(0)), numeric(0))
-  expect_identical(qburrIII2(0), 0)
-  expect_identical(qburrIII2(1), Inf)
-  expect_identical(qburrIII2(NA), NA_real_)
-  expect_identical(qburrIII2(0.5, lscale = 0), 1)
-  expect_equal(qburrIII2(c(0.1, 0.2), lscale = 0), c(0.111111111111111, 0.25))
+test_that("burrIII2 and llogis identical", {
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
+  burr <- ssd_fit_dist(ssdtools::boron_data, dist = "burrIII2")
+  llogis <- ssd_fit_dist(ssdtools::boron_data, dist = "llogis")
+  expect_identical(ssd_hc(burr)$est, ssd_hc(llogis)$est)
 })
 
-test_that("pburrIII2", {
-  expect_identical(pburrIII2(numeric(0)), numeric(0))
-  expect_identical(pburrIII2(0), 0)
-  expect_identical(pburrIII2(1, lscale = 0), 0.5)
-  expect_identical(pburrIII2(NA), NA_real_)
-  expect_identical(pburrIII2(qburrIII2(0.5)), 0.5)
-  expect_equal(
-    pburrIII2(qburrIII2(c(.001, .01, .05, .50, .99))),
-    c(.001, .01, .05, .50, .99)
-  )
+test_that("deprecated dburrIII2", {
+  expect_error(ssd_fit_dists(ssdtools::boron_data, dist = c("llog", "burrIII2")), "Distributions 'llog', 'burrIII2' and 'llogis' are identical. Please just use 'llogis'.")
 })
 
-test_that("rburrIII2", {
-  expect_identical(rburrIII2(0), numeric(0))
-  set.seed(101)
-  expect_equal(
-    rburrIII2(10, lscale = 0),
-    c(
-      0.592859849849427, 0.0458334582732346, 2.44452273715775, 1.92133200433186,
-      0.33307689063302, 0.428683341542249, 1.40886438557549, 0.500301132990527,
-      1.6455863786175, 1.20181169357475
-    )
-  )
+test_that("deprecated dburrIII2", {
+  expect_error(ssd_fit_dists(ssdtools::boron_data, dist = c("llogis", "burrIII2")), "Distributions 'llog', 'burrIII2' and 'llogis' are identical. Please just use 'llogis'.")
 })
+
+