+++ tests/testthat/test-pareto.R
@@ -13,6 +13,8 @@
 #    limitations under the License.
 
 test_that("dpareto", {
+  lifecycle::expect_deprecated(dpareto(1))
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
   expect_equal(dpareto(numeric(0)), numeric(0))
   expect_equal(dpareto(c(0, 1, Inf, NaN, -1)), c(0, 1, 0, NA, 0))
   expect_equal(dpareto(2), 0.25)
@@ -23,7 +25,8 @@ test_that("dpareto", {
 })
 
 test_that("fit pareto", {
-  dist <- ssdtools:::ssd_fit_dist(ssdtools::boron_data, dist = "pareto")
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
+  dist <- ssd_fit_dist(ssdtools::boron_data, dist = "pareto")
 
   expect_true(is.fitdist(dist))
   expect_equal(
@@ -33,8 +36,11 @@ test_that("fit pareto", {
 })
 
 test_that("fit pareto cis", {
-  dist <- ssdtools:::ssd_fit_dist(ssdtools::boron_data, dist = "pareto")
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
+  dist <- ssd_fit_dist(ssdtools::boron_data, dist = "pareto")
 
+  skip_if_not(capabilities("long.double"))
+  
   set.seed(77)
   expect_equal(
     as.data.frame(ssd_hc(dist, ci = TRUE, nboot = 10)),
@@ -44,9 +50,7 @@ test_that("fit pareto cis", {
     ), class = "data.frame", row.names = c(
       NA,
       -1L
-    )),
-    tolerance = 1e-07 # for noLD
-  )
+    )))
   set.seed(77)
   expect_equal(
     as.data.frame(ssd_hp(dist, conc = 2, ci = TRUE, nboot = 10)),
@@ -56,12 +60,12 @@ test_that("fit pareto cis", {
     ), class = "data.frame", row.names = c(
       NA,
       -1L
-    )),
-    tolerance = 1e-07 # for noLD
-  )
+    )))
 })
 
 test_that("ppareto", {
+  lifecycle::expect_deprecated(ppareto(1))
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
   expect_equal(ppareto(1), 0)
   expect_equal(ppareto(2), 1 / 2)
   expect_equal(ppareto(3), 2 / 3)
@@ -71,11 +75,15 @@ test_that("ppareto", {
 })
 
 test_that("qpareto", {
+  lifecycle::expect_deprecated(qpareto(1))
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
   expect_equal(qpareto(ppareto(2)), 2)
   expect_equal(qpareto(numeric(0)), numeric(0))
 })
 
 test_that("rpareto", {
+  lifecycle::expect_deprecated(rpareto(1))
+  rlang::scoped_options(lifecycle_verbosity = "quiet")
   set.seed(1)
   expect_equal(rpareto(1), 3.7663554483147)
 })