+++ NEWS.md
@@ -1,3 +1,28 @@
+# ssdtools 0.3.0
+
+## Breaking Changes
+
+- Soft-deprecated 'burrIII3' distribution as poorly defined.
+- Soft-deprecated 'pareto' distribution as poor fit on SSD data.
+
+## Major Changes
+
+- Reparameterized 'llogis' distribution in terms of locationlog and scalelog.
+- Reparameterized 'burrIII3' distribution in terms of lshape1, lshape2 and lscale.
+- Reparamaterized 'burrIII2' distribution in terms of locationlog and scalelog.
+- Reparamaterized 'lgumbel' distribution in terms of locationlog and scalelog.
+- Reparamaterized 'gompertz' distribution in terms of llocation and lshape.
+- Standardized handling of arguments for d,p,q,r and s functions for distributions.
+
+## Minor Changes
+
+- `rdist()` functions now use length of n if `length(n) > 1`.
+- Added `slnorm()` to get starting values for 'dlnorm' distribution.
+
+## Internal Changes
+
+- Switch to C++ implementation for distributions.
+
 # ssdtools 0.2.0
 
 ## Breaking Changes